[Core] Font: Remove UseCase to get TextStyle
